ENH: try using uint16_t instead of uint8_t for kmer words - uint8_t potentially is slower for MPI.

+++ test/test/compact_debruijn_graph_block_construct.cpp
@@ -104,9 +104,9 @@ using Alphabet = bliss::common::DNA;
 
 
 #if defined(pK)
-using KmerType = bliss::common::Kmer<pK, Alphabet, uint8_t>;
+using KmerType = bliss::common::Kmer<pK, Alphabet, uint16_t>;
 #else
-using KmerType = bliss::common::Kmer<31, Alphabet, uint8_t>;
+using KmerType = bliss::common::Kmer<31, Alphabet, uint16_t>;
 #endif
